Extract the Skeleton component
