+++ packages/template-retail-react-app/app/components/product-tile/skeleton.jsx
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2023, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+import React from 'react'
+import {
+    useMultiStyleConfig,
+    AspectRatio,
+    Box,
+    Stack,
+    Skeleton as ChakraSkeleton
+} from '@salesforce/retail-react-app/app/components/shared/ui'
+
+// TODO: remove this no-longer-used skeleton
+const Skeleton = () => {
+    const styles = useMultiStyleConfig('ProductTile')
+    return (
+        <Box data-testid="sf-product-tile-skeleton">
+            <Stack spacing={2}>
+                <Box {...styles.imageWrapper}>
+                    <AspectRatio ratio={1} {...styles.image}>
+                        <ChakraSkeleton />
+                    </AspectRatio>
+                </Box>
+                <ChakraSkeleton width="80px" height="20px" />
+                <ChakraSkeleton width={{base: '120px', md: '220px'}} height="12px" />
+            </Stack>
+        </Box>
+    )
+}
+export default Skeleton
